fix: Create Invoice error — decimal(10,2) not supported in SQLite
- Changed all monetary fields from decimal to double (SQLite stores REAL)
- Added [NotMapped] to InvoiceItem.Total computed property so EF Core does not try to map it
- Fixed CREATE TABLE SQL to use REAL instead of TEXT for amount columns

+++ src/HealthcareCRM/Controllers/Api/BillingController.cs
@@ -229,14 +229,14 @@ namespace HealthcareCRM.Controllers.Api
         [System.ComponentModel.DataAnnotations.Required]
         public string Description { get; set; } = string.Empty;
         [System.ComponentModel.DataAnnotations.Required]
-        public decimal UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
         public int Quantity { get; set; } = 1;
     }
 
     public class RecordPaymentRequest
     {
         [System.ComponentModel.DataAnnotations.Required]
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
         public string? Method { get; set; } = "Cash";
         public string? Reference { get; set; }
         public string? Notes { get; set; }
